pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "maxtexlib"
version = "0.1"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  { name = "Max Lindqvist"},
  { email = "max.lindqvist@ipp.mpg.de"}
]
description = "A simple template for generating reports with LaTeX and figures with maxplotlib"
keywords = ["LaTeX, template"]

classifiers = [
    "Programming Language :: Python :: 3",
    ]

dependencies = [
    'pyyaml',
    'maxtikzlib @ git+https://github.com/max-models/maxtikzlib.git@main',
    'maxplotlib @ git+https://github.com/max-models/maxplotlib.git@main',
]


[project.urls]
# homepage = ""
# documentation = ""
repository = "https://gitlab.mpcdf.mpg.de/maxlin/template-latex"
# changelog = ""
# "Bug Tracker" = ""

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
maxtexlib = "maxtexlib.main:main"

[tool.setuptools.package-data]
"maxtexlib.templates" = ["*.tex", "makefile", "tex/*.tex", "figures/*.py"]

[tool.ruff]
line-length=120
